+++ jabgui/src/main/java/org/jabref/gui/frame/MainMenu.java
@@ -376,6 +376,13 @@ public class MainMenu extends MenuBar {
 
                 new SeparatorMenuItem(),
 
+                // [impl->req~ux.background-tasks.viewer~1]
+                factory.createMenuItem(StandardActions.SHOW_BACKGROUND_TASKS, new SimpleCommand() {
+                    @Override
+                    public void execute() {
+                        dialogService.showBackgroundTasksDialog(stateManager);
+                    }
+                }),
                 factory.createMenuItem(StandardActions.ERROR_CONSOLE, new ErrorConsoleAction()),
 
                 new SeparatorMenuItem(),